+++ cluster-operator/src/test/java/io/strimzi/operator/cluster/operator/resource/KafkaBrokerConfigurationDiffTest.java
@@ -33,7 +33,7 @@ import static org.junit.jupiter.api.Assertions.fail;
 public class KafkaBrokerConfigurationDiffTest {
 
     private static final KafkaVersion.Lookup VERSIONS = KafkaVersionTestUtils.getKafkaVersionLookup();
-    private static final String KAFKA_VERSION = "3.5.1";
+    private static final String KAFKA_VERSION = "3.6.0";
     KafkaVersion kafkaVersion = VERSIONS.supportedVersion(KAFKA_VERSION);
     private final int brokerId = 0;